requirements.txt
opencv-python==4.11.0.86
numpy==1.26.4
GitPython==3.1.44
gitdb==4.0.12
Pillow==9.5.0
pip==23.2.1
attrs==25.3.0
protobuf==4.25.7
tornado==6.4.2
Jinja2==3.1.6
pytz==2025.2
rich==13.9.4
Pygments==2.19.1
markdown-it-py==3.0.0
toml==0.10.2
zipp==3.21.0
typing_extensions==4.13.2
smmap==5.0.2
pywin32==310
mdurl==0.1.2
charset-normalizer==3.4.2
altair==5.5.0
jsonschema==4.23.0
narwhals==1.37.1
pandas==2.2.3
pyarrow==20.0.0
packaging==23.2
referencing==0.36.2
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
pydeck==0.9.1
certifi==2025.4.26
tzdata==2025.2
Pympler==1.1
colorama==0.4.6
tzlocal==4.3.1
urllib3==2.4.0
six==1.17.0
idna==3.10
tenacity==8.5.0
watchdog==6.0.0
click==8.1.8
streamlit==1.25.0
cachetools==5.5.2
blinker==1.9.0
importlib-metadata==6.11.0
jsonschema-specifications==2025.4.1
validators==0.35.0